Add: Enclose the PerformOrder inside a class
